Preserve category extra values when form fields are reloaded
Any category extra input was being lost when logging in during report
creation. This commit re-sets the values of any form inputs when the
category extra elements are reloaded, e.g. when the category is changed.

I’m not entirely sure this is the right solution, and it doesn’t seem to
work in all cases (e.g. clicking elsewhere on the map when creating a
report), but it fixes the login bug for now.